# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TripSync API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trips:
    get:
      operationId: listTrips
      tags: [trips]
      summary: List all trips
      responses:
        "200":
          description: List of trips
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trip"
    post:
      operationId: createTrip
      tags: [trips]
      summary: Create a new trip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripInput"
      responses:
        "201":
          description: Created trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
  /trips/{id}:
    get:
      operationId: getTrip
      tags: [trips]
      summary: Get a trip by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trip details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
        "404":
          description: Not found
    patch:
      operationId: updateTrip
      tags: [trips]
      summary: Update a trip
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripUpdate"
      responses:
        "200":
          description: Updated trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
        "404":
          description: Not found
    delete:
      operationId: deleteTrip
      tags: [trips]
      summary: Delete a trip
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /trips/{id}/summary:
    get:
      operationId: getTripSummary
      tags: [analytics]
      summary: Get trip dashboard summary
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trip summary stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripSummary"
        "404":
          description: Not found
  /trips/{id}/expense-analytics:
    get:
      operationId: getTripExpenseAnalytics
      tags: [analytics]
      summary: Get expense analytics for a trip
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Expense analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseAnalytics"
        "404":
          description: Not found

  # ─── MEMBERS ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Trip:
      type: object
      required: [id, name, destination, startDate, endDate, budget, currency, status, description, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        destination:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        budget:
          type: number
        currency:
          type: string
        status:
          type: string
          enum: [planning, active, completed]
        description:
          type: ["string", "null"]
        coverImage:
          type: ["string", "null"]
        createdAt:
          type: string
    TripInput:
      type: object
      required: [name, destination, startDate, endDate, budget, currency]
      properties:
        name:
          type: string
          minLength: 1
        destination:
          type: string
          minLength: 1
        startDate:
          type: string
        endDate:
          type: string
        budget:
          type: number
        currency:
          type: string
        description:
          type: string
        coverImage:
          type: string
    TripUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        destination:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        budget:
          type: number
        currency:
          type: string
        status:
          type: string
          enum: [planning, active, completed]
        description:
          type: string
        coverImage:
          type: string

    # ── TRIP SUMMARY (wow endpoint) ───────────────────────────────────────────
    TripSummary:
      type: object
      required: [tripId, totalBudget, totalSpent, remaining, budgetUsedPct, memberCount, itineraryCount, taskCount, pendingTaskCount, expenseCount]
      properties:
        tripId:
          type: integer
        totalBudget:
          type: number
        totalSpent:
          type: number
        remaining:
          type: number
        budgetUsedPct:
          type: number
        memberCount:
          type: integer
        itineraryCount:
          type: integer
        taskCount:
          type: integer
        pendingTaskCount:
          type: integer
        expenseCount:
          type: integer

    # ── EXPENSE ANALYTICS (wow endpoint) ─────────────────────────────────────
    ExpenseAnalytics:
      type: object
      required: [tripId, totalSpent, byCategory, byPerson]
      properties:
        tripId:
          type: integer
        totalSpent:
          type: number
        byCategory:
          type: array
          items:
            $ref: "#/components/schemas/CategoryBreakdown"
        byPerson:
          type: array
          items:
            $ref: "#/components/schemas/PersonSpending"
    CategoryBreakdown:
      type: object
      required: [category, amount, count]
      properties:
        category:
          type: string
        amount:
          type: number
        count:
          type: integer
    PersonSpending:
      type: object
      required: [name, paid, owes]
      properties:
        name:
          type: string
        paid:
          type: number
        owes:
          type: number
...