# lib/api-zod/src/generated/types/expenseAnalytics.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TripSync API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryBreakdown } from "./categoryBreakdown";
import type { PersonSpending } from "./personSpending";

export interface ExpenseAnalytics {
  tripId: number;
  totalSpent: number;
  byCategory: CategoryBreakdown[];
  byPerson: PersonSpending[];
}