# lib/api-zod/src/generated/types/tripUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TripSync API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TripUpdateStatus } from "./tripUpdateStatus";

export interface TripUpdate {
  /** @minLength 1 */
  name?: string;
  destination?: string;
  startDate?: string;
  endDate?: string;
  budget?: number;
  currency?: string;
  status?: TripUpdateStatus;
  description?: string;
  coverImage?: string;
}